Replace 'inline' with 'static inline'
This change enables compilers to correctly scope the symbol and apply
inlining optimizations. For example, clang emits functions and calls in
the code, which is suboptimal. By assisting the compiler, we ensure that
this function, not used elsewhere, does not need global scope. Without
this modification, the project cannot be built on macOS, where Clang is
the default compiler.